requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

pydantic==2.10.3
pydantic-settings==2.6.1

SQLAlchemy==2.0.36
asyncpg==0.30.0

httpx==0.28.1

pandas==2.2.3
numpy==2.1.3

plotly==5.24.1

python-dotenv==1.0.1

# IA (opcional)
langchain==0.3.13
langchain-core==0.3.27
langchain-openai==0.2.14
google-generativeai==0.8.3